Updating template generation
