+++ crates/pop-cli/src/commands/new/pallet.rs
@@ -2,16 +2,14 @@
 
 use crate::{
 	cli::{traits::Cli as _, Cli},
-	common::helpers::{collect_loop_cliclack_inputs, valid_ident},
 };
 
-use crate::{multiselect_pick, pick_options_and_give_name};
+use pop_common::multiselect_pick;
 use clap::{Args, Subcommand};
-use cliclack::{confirm, input, multiselect, outro, outro_cancel, select};
+use cliclack::{confirm, multiselect, outro, outro_cancel};
 use pop_parachains::{
 	create_pallet_template, resolve_pallet_path, TemplatePalletConfig,
-	TemplatePalletConfigCommonTypes, TemplatePalletConfigTypesDefault,
-	TemplatePalletConfigTypesMetadata, TemplatePalletStorageTypes,
+	TemplatePalletConfigCommonTypes, TemplatePalletStorageTypes
 };
 use std::fs;
 use strum::{EnumMessage, IntoEnumIterator};
@@ -57,11 +55,9 @@ impl NewPalletCommand {
 
 		let mut pallet_default_config = false;
 		let mut pallet_common_types = Vec::new();
-		let mut pallet_config_types = Vec::new();
 		let mut pallet_storage = Vec::new();
 		let mut pallet_genesis = false;
 		let mut pallet_custom_origin = false;
-		let mut pallet_custom_origin_variants = Vec::new();
 
 		if let Some(Mode::Advanced(advanced_mode_args)) = &self.mode {
 			pallet_default_config = advanced_mode_args.default_config;
@@ -73,43 +69,14 @@ impl NewPalletCommand {
 
 				pallet_common_types = multiselect_pick!(TemplatePalletConfigCommonTypes, "Are you interested in adding one of these types and their usual configuration to your pallet?");
 
-				Cli.info("If you want to add some other config types, this is the moment. Keep adding them until you're done!")?;
-
-				// Depending on the user's selection, the cli should offer to choose wheter the type is
-				// included in the default config or not.
-				if advanced_mode_args.default_config {
-					pallet_config_types = pick_options_and_give_name!(
-                        (TemplatePalletConfigTypesMetadata ,"You're adding a new type to your pallet's Config trait! Should it be included in the metadata?"), 
-                        (TemplatePalletConfigTypesDefault, "Should it be included in the default configuration?")
-                    );
-				} else {
-					// As the vec must contain as well info related to the default config, we map everything
-					// to have the Default variant. The type annotations happen when the type's no default,
-					// so this variant ensures there's no type annotation related to default, even though
-					// there's no default config and then the types cannot be default
-					pallet_config_types = pick_options_and_give_name!(
-                        (TemplatePalletConfigTypesMetadata ,"You're adding a new type to your pallet's Config trait! Should it be included in the metadata?")
-                    )
-                        .into_iter()
-                        .map(|(to_metadata, config_type)| (to_metadata, TemplatePalletConfigTypesDefault::Default, config_type))
-                        .collect::<Vec<(TemplatePalletConfigTypesMetadata, TemplatePalletConfigTypesDefault, String)>>();
-				}
 			}
 
 			if advanced_mode_args.storage {
 				Cli.info("Generate the pallet's storage.")?;
 
-				pallet_storage = pick_options_and_give_name!(
-                    (TemplatePalletStorageTypes,"You're adding a new storage item to your pallet! Select a storage type to create an instance of it:")
-                );
+				pallet_storage = multiselect_pick!(TemplatePalletStorageTypes,"Are you interested in adding some of those storage items to your pallet?");
 			}
 
-			if pallet_custom_origin {
-				Cli.info("Generate the pallet's custom origin")?;
-
-				pallet_custom_origin_variants =
-					collect_loop_cliclack_inputs("Add a variant name.")?;
-			}
 		};
 
 		let target = resolve_pallet_path(self.path.clone())?;
@@ -141,11 +108,10 @@ impl NewPalletCommand {
 				pallet_advanced_mode: self.mode.is_some(),
 				pallet_default_config,
 				pallet_common_types,
-				pallet_config_types,
 				pallet_storage,
 				pallet_genesis,
 				pallet_custom_origin,
-				pallet_custom_origin_variants,
+
 			},
 		)?;
