--- crates/pop-dev/README.md
@@ -1,1 +1,0 @@
-A small crate to help with POP development
