--- crates/pop-dev/src/lib.rs
@@ -1,3 +1,0 @@
-mod placeholders;
-
-pub use placeholders::*;
