+++ crates/pop-parachains/src/lib.rs
@@ -18,10 +18,7 @@ pub use errors::Error;
 pub use indexmap::IndexSet;
 pub use new_pallet::{
 	create_pallet_template,
-	new_pallet_options::{
-		TemplatePalletConfigCommonTypes, TemplatePalletConfigTypesDefault,
-		TemplatePalletConfigTypesMetadata, TemplatePalletStorageTypes,
-	},
+	new_pallet_options::*,
 	TemplatePalletConfig,
 };
 pub use new_parachain::instantiate_template_dir;
