+++ crates/pop-parachains/templates/pallet/advanced_mode/src/config_preludes.rs.templ
@@ -3,7 +3,6 @@
 {{- let using_currency = pallet_common_types|contains(TemplatePalletConfigCommonTypes::Currency) -}}
 use super::*;
 use frame::runtime::prelude::derive_impl;
-use pop_dev::{PlaceholderType, pop_todo};
 
 /// Provides a viable default config that can be used with
 /// [`derive_impl`] to derive a testing pallet config
@@ -29,25 +28,6 @@ impl DefaultConfig for TestDefaultConfig {
     #[inject_runtime_type]
     type RuntimeFreezeReason = ();
     {{- endif -}}
-    {{- for config_type in pallet_config_types -}}
-    {{- match config_type.1 -}}
-    {{- when TemplatePalletConfigTypesDefault::Default }}
-    {{- match config_type.0 -}}
-    {{- when TemplatePalletConfigTypesMetadata::Constant }}
-    type ^^config_type.2|capitalize^^ = pop_todo!(constant_default_type);
-    {{- else }}
-    type ^^config_type.2|capitalize^^ = PlaceholderType;
-    {{- endmatch -}}
-    {{- when TemplatePalletConfigTypesDefault::NoDefaultBounds }}
-    {{- match config_type.0 -}}
-    {{- when TemplatePalletConfigTypesMetadata::Constant }}
-    type ^^config_type.2|capitalize^^ = pop_todo!(constant_default_type);
-    {{- else }}
-    type ^^config_type.2|capitalize^^ = PlaceholderType;
-    {{- endmatch -}}
-    {{- else -}}
-    {{- endmatch -}}
-    {{- endfor }}
 }
 
 /// Default configurations of this pallet in a solochain environment.
@@ -72,25 +52,6 @@ impl DefaultConfig for SolochainDefaultConfig {
     #[inject_runtime_type]
     type RuntimeFreezeReason = ();
     {{- endif -}}
-    {{- for config_type in pallet_config_types -}}
-    {{- match config_type.1 -}}
-    {{- when TemplatePalletConfigTypesDefault::Default }}
-    {{- match config_type.0 -}}
-    {{- when TemplatePalletConfigTypesMetadata::Constant }}
-    type ^^config_type.2|capitalize^^ = pop_todo!(constant_default_type);
-    {{- else }}
-    type ^^config_type.2|capitalize^^ = PlaceholderType;
-    {{- endmatch -}}
-    {{- when TemplatePalletConfigTypesDefault::NoDefaultBounds }}
-    {{- match config_type.0 -}}
-    {{- when TemplatePalletConfigTypesMetadata::Constant }}
-    type ^^config_type.2|capitalize^^ = pop_todo!(constant_default_type);
-    {{- else }}
-    type ^^config_type.2|capitalize^^ = PlaceholderType;
-    {{- endmatch -}}
-    {{- else -}}
-    {{- endmatch -}}
-    {{- endfor }}
 }
 
 /// Default configurations of this pallet in a solochain environment.
@@ -115,25 +76,6 @@ impl DefaultConfig for RelayChainDefaultConfig {
     #[inject_runtime_type]
     type RuntimeFreezeReason = ();
     {{- endif -}}
-    {{- for config_type in pallet_config_types -}}
-    {{- match config_type.1 -}}
-    {{- when TemplatePalletConfigTypesDefault::Default }}
-    {{- match config_type.0 -}}
-    {{- when TemplatePalletConfigTypesMetadata::Constant }}
-    type ^^config_type.2|capitalize^^ = pop_todo!(constant_default_type);
-    {{- else }}
-    type ^^config_type.2|capitalize^^ = PlaceholderType;
-    {{- endmatch -}}
-    {{- when TemplatePalletConfigTypesDefault::NoDefaultBounds }}
-    {{- match config_type.0 -}}
-    {{- when TemplatePalletConfigTypesMetadata::Constant }}
-    type ^^config_type.2|capitalize^^ = pop_todo!(constant_default_type);
-    {{- else }}
-    type ^^config_type.2|capitalize^^ = PlaceholderType;
-    {{- endmatch -}}
-    {{- else -}}
-    {{- endmatch -}}
-    {{- endfor }}
 } 
 
 /// Default configurations of this pallet in a solochain environment.
@@ -158,23 +100,4 @@ impl DefaultConfig for ParaChainDefaultConfig{
     #[inject_runtime_type]
     type RuntimeFreezeReason = ();
     {{- endif -}}
-    {{- for config_type in pallet_config_types -}}
-    {{- match config_type.1 -}}
-    {{- when TemplatePalletConfigTypesDefault::Default }}
-    {{- match config_type.0 -}}
-    {{- when TemplatePalletConfigTypesMetadata::Constant }}
-    type ^^config_type.2|capitalize^^ = pop_todo!(constant_default_type);
-    {{- else }}
-    type ^^config_type.2|capitalize^^ = PlaceholderType;
-    {{- endmatch -}}
-    {{- when TemplatePalletConfigTypesDefault::NoDefaultBounds }}
-    {{- match config_type.0 -}}
-    {{- when TemplatePalletConfigTypesMetadata::Constant }}
-    type ^^config_type.2|capitalize^^ = pop_todo!(constant_default_type);
-    {{- else }}
-    type ^^config_type.2|capitalize^^ = PlaceholderType;
-    {{- endmatch -}}
-    {{- else -}}
-    {{- endmatch -}}
-    {{- endfor }}
 }
