+++ crates/pop-parachains/templates/pallet/advanced_mode/src/lib.rs.templ
@@ -10,8 +10,6 @@
 
 use frame::prelude::*;
 
-use pop_dev::{PlaceholderTraitBound, pop_todo};
-
 {{ if using_currency }}
 use frame::traits::{fungible, VariantCount};
 {{- endif }}
@@ -109,65 +107,71 @@ pub mod pallet {
 			+ fungible::freeze::Mutate<Self::AccountId>;
         {{- endif }}
 
-        {{ for config_type in pallet_config_types -}}
-        /// Some docs for this type
-        {{- match config_type.1 -}}
-        {{- when TemplatePalletConfigTypesDefault::NoDefaultBounds }}
-        #[pallet::no_default_bounds]
-        {{- when TemplatePalletConfigTypesDefault::NoDefault }}
-        #[pallet::no_default]
-        {{- else -}}
-        {{- endmatch }}
-        {{- match config_type.0 -}}
-        {{- when TemplatePalletConfigTypesMetadata::Constant }}
-        #[pallet::constant]
-        type ^^config_type.2|capitalize^^: Get<u32>;
-
-        {{ else }}
-        type ^^config_type.2|capitalize^^: PlaceholderTraitBound;
-
-        {{ endmatch -}}
-        {{ endfor }}
     }
 
-    {{ for storage in pallet_storage }}
-    /// Some docs for this storage type
+    {{ if pallet_storage|contains(TemplatePalletStorageTypes::StorageMap) }}
+    /// A fancy struct that can be used to store information together in storage. It's storing a phrase chosen by a AccountId and the Block number when it's registered SCALE compact encoded
+    #[derive(TypeInfo, Encode, Decode, MaxEncodedLen, PartialEq, RuntimeDebugNoBound)]
+	#[scale_info(skip_type_params(T))]
+	pub struct AccountInfo<T: Config> {
+        // An awesome phrase 
+		pub phrase: String,
+        // The block where the account registered the phrase
+		#[codec(compact)]
+		pub block_joined: BalanceOf<T>,
+	}
+    {{ endif }}
+
+    {{ if pallet_storage|contains(TemplatePalletStorageTypes::StorageValue) }}
+    /// A simple StorageValue, storing a u32
     #[pallet::storage]
-    {{- match storage.0 -}}
-    {{- when TemplatePalletStorageTypes::StorageValue }}
-    pub type ^^storage.1|capitalize^^<T: Config> = StorageValue<Value = pop_todo!(storage_value)>;
+    pub type MyStorageValue = StorageValue<Value = u32>;
+    {{ endif }}
 
-    {{ when TemplatePalletStorageTypes::StorageMap }}
-    pub type ^^storage.1|capitalize^^<T: Config> = StorageMap<Hasher = pop_todo!(storage_hasher), Key = pop_todo!(storage_key), Value = pop_todo!(storage_value)>;
+    {{ if pallet_storage|contains(TemplatePalletStorageTypes::StorageMap) }}
+    /// A simple StorageMap, mapping AccountIds with AccountsInfo! Using Twox64Concat as Hasher is Ok cause T::AccountId is generated by a cryptographically secure hash function, so the keys won't create an unbalanced trie
+    #[pallet::storage]
+    pub type MyStorageMap<T:Config> = StorageMap<Hasher = Twox64Concat, Key: T::AccountId, Value = AccountInfo<T>>;
+    {{ endif }}
 
-    {{ when TemplatePalletStorageTypes::CountedStorageMap }}
-    pub type ^^storage.1|capitalize^^<T: Config> = CountedStorageMap<Hasher = pop_todo!(storage_hasher), Key = pop_todo!(storage_key), Value = pop_todo!(storage_value)>;
+    {{ if pallet_storage|contains(TemplatePalletStorageTypes::CountedStorageMap) }}
+    /// A CountedStorageMap mapping u32 with some AccountIds. Blake2_128Concat is needed as Hasher to store the keys in a cryptographically secure way in order to create a balanced trie
+    #[pallet::storage]
+    pub type MyCountedStorageMap<T:Config> = CountedStorageMap<Hasher = Blake2_128Concat, Key: u32, Value = T::AccountId>;
+    {{ endif }}
 
-    {{ when TemplatePalletStorageTypes::StorageDoubleMap }}
-    pub type ^^storage.1|capitalize^^<T: Config> = StorageDoubleMap<Hasher1 = pop_todo!(storage_hasher), Key1 = pop_todo!(storage_key), Hasher2 = pop_todo!(storage_hasher), Key2 = pop_todo!(storage_key), Value = pop_todo!(storage_value)>;
+    {{ if pallet_storage|contains(TemplatePalletStorageTypes::StorageDoubleMap) }}
+    /// A StorageDoubleMap mapping AccountIds and Block Numbers to a value. The first hasher can be Twox64Concat as T::AccountId is generated by a cryptographically secure hash function, so the keys won't create an unbalanced trie, Blake2_128Concat is needed as Hasher2 to store the keys in a cryptographically secure way in order to create a balanced trie
+    #[pallet::storage]
+    pub type MyStorageDoubleMap<T:Config> = StorageDoubleMap<Hasher1 = Twox64Concat, Key1 = T::AccountId, Hasher2 = Blake2_128Concat, Key2 = BlockNumberFor<T>, Value = u32>;
+    {{ endif }}
 
-    {{ when TemplatePalletStorageTypes::StorageNMap }}
-    pub type ^^storage.1|capitalize^^<T: Config>= StorageNMap<
+    {{ if pallet_storage|contains(TemplatePalletStorageTypes::StorageNMap) }}
+    /// A StorageNMap mapping 3 AccountIds to a u32 value, all the hashers are Twox64Concat as T::AccountId is generated by a cryptographically secure hash function, so the keys won't create an unbalanced trie.
+    #[pallet::storage]
+    pub type MyStorageNMap<T: Config>= StorageNMap<
         Key = (
-            NMapKey<pop_todo!(storage_hasher), pop_todo!(storage_key)>,
-            NMapKey<pop_todo!(storage_hasher), pop_todo!(storage_key)>,
-            NMapKey<pop_todo!(storage_hasher), pop_todo!(storage_key)>
+            NMapKey<Twox64Concat, T::AccountId>,
+            NMapKey<Twox64Concat, T::AccountId>,
+            NMapKey<Twox64Concat, T::AccountId>
         ),
-        Value = pop_todo!(storage_value)
+        Value = u32
     >;
+    {{ endif }}
 
-    {{ when TemplatePalletStorageTypes::CountedStorageNMap }}
-    pub type ^^storage.1|capitalize^^<T: Config> = CountedStorageNMap<
+    {{ if pallet_storage|contains(TemplatePalletStorageTypes::StorageNMap) }}
+    /// A CountedStorageNMap mapping 3 AccountIds to a u32 value, all the hashers are Twox64Concat as T::AccountId is generated by a cryptographically secure hash function, so the keys won't create an unbalanced trie.
+    #[pallet::storage]
+    pub type MyCountedStorageNMap<T: Config>= CountedStorageNMap<
         Key = (
-            NMapKey<pop_todo!(storage_hasher), pop_todo!(storage_key)>,
-            NMapKey<pop_todo!(storage_hasher), pop_todo!(storage_key)>,
-            NMapKey<pop_todo!(storage_hasher), pop_todo!(storage_key)>
+            NMapKey<Twox64Concat, T::AccountId>,
+            NMapKey<Twox64Concat, T::AccountId>,
+            NMapKey<Twox64Concat, T::AccountId>
         ),
-        Value = pop_todo!(storage_value)
+        Value = u32
     >;
-    
-    {{ endmatch -}}
-    {{- endfor -}}
+    {{ endif }}
+
 
     {{ if using_runtime_event }}
 	#[pallet::event]
@@ -207,12 +211,20 @@ pub mod pallet {
     {{ endif -}}
 
     {{- if pallet_custom_origin }}
+    /// Custom origin for this pallet
     #[pallet::origin]
     #[derive(PartialEq, Eq, Clone, RuntimeDebug, Encode, Decode, TypeInfo, MaxEncodedLen)]
-    pub enum Origin {
-        {{- for variant in pallet_custom_origin_variants }}
-            ^^variant|capitalize^^,
-        {{- endfor -}}
+    pub enum Origin<T: Config> {
+        /// This origin's only valid if the current block number is greater than a certain block
+        BlockNumberPassed(BlockNumberFor<T>),
+        {{ if pallet_storage|contains(TemplatePalletStorageTypes::StorageValue) }}
+        /// This origin's only valid if MyStorageValue has been set
+        StorageValueNotEmpty,
+        {{ endif }}
+        {{ if pallet_storage|contains(TemplatePalletStorageTypes::StorageMap) }}
+        /// This origin's only valid if the inner AccountId is present in MyStorageMap
+        AccountWithInfo(T::AccountId),
+        {{ endif }}
     }
 
     {{ endif -}}
