+++ crates/pop-parachains/templates/pallet/advanced_mode/src/mock.rs.templ
@@ -2,7 +2,6 @@
 {{- let using_runtime_origin = pallet_common_types|contains(TemplatePalletConfigCommonTypes::RuntimeOrigin) -}}
 {{- let using_currency = pallet_common_types|contains(TemplatePalletConfigCommonTypes::Currency) -}}
 use crate::{self as pallet_^^name|lower^^, *};
-use pop_dev::{PlaceholderType, pop_todo};
 
 use frame::{
 	deps::frame_support::runtime,
@@ -80,35 +79,6 @@ impl pallet_^^name|lower^^::Config for Test {
     type RuntimeFreezeReason = RuntimeFreezeReason;
     {{- endif -}}
     {{- endif -}}
-    {{- for config_type in pallet_config_types -}}
-    {{- match config_type.1 -}}
-    {{- when TemplatePalletConfigTypesDefault::Default -}}
-    {{- if !pallet_default_config -}}
-    {{- match config_type.0 -}}
-    {{- when TemplatePalletConfigTypesMetadata::Constant }}
-    type ^^config_type.2|capitalize^^ = pop_todo!(constant_default_type);
-    {{- else }}
-    type ^^config_type.2|capitalize^^ = PlaceholderType;
-    {{- endmatch -}}
-    {{- endif -}}
-    {{- when TemplatePalletConfigTypesDefault::NoDefaultBounds }}
-    {{- if !pallet_default_config }}
-    {{- match config_type.0 -}}
-    {{- when TemplatePalletConfigTypesMetadata::Constant }}
-    type ^^config_type.2|capitalize^^ = pop_todo!(constant_default_type);
-    {{- else }}
-    type ^^config_type.2|capitalize^^ = PlaceholderType;
-    {{- endmatch -}}
-    {{- endif -}}
-    {{- else }}
-    {{- match config_type.0 -}}
-    {{- when TemplatePalletConfigTypesMetadata::Constant }}
-    type ^^config_type.2|capitalize^^ = pop_todo!(constant_default_type);
-    {{- else }}
-    type ^^config_type.2|capitalize^^ = PlaceholderType;
-    {{- endmatch -}}
-    {{- endmatch -}}
-    {{- endfor -}}
 }
 
 pub(crate) struct StateBuilder{
